ProtocolGen: use size_t to work with type sizes

+++ src/private/protocolgen/cpphelper.cpp
@@ -31,7 +31,7 @@ namespace {
 
 class Dummy;
 
-static QHash<QByteArray, int> typeSizeLookup = {
+static QHash<QByteArray, size_t> typeSizeLookup = {
     { "Scope",        sizeof(QSharedDataPointer<Dummy>) },
     { "ScopeContext", sizeof(QSharedDataPointer<Dummy>) },
     { "QSharedPointer", sizeof(QSharedPointer<Dummy>) },
@@ -45,7 +45,7 @@ static QHash<QByteArray, int> typeSizeLookup = {
 // and does not consider alignment. It should be good enough (TM) for our needs,
 // but it would be nice to make it smarter, for example by looking up type sizes
 // from the Node tree and understanding enums and QFlags types.
-int typeSize(const QString &typeName)
+size_t typeSize(const QString &typeName)
 {
     QByteArray tn;
     // Don't you just loooove hacks?
@@ -63,7 +63,7 @@ int typeSize(const QString &typeName)
         const auto typeId = QMetaType::type(tn);
         const int size = QMetaType(typeId).sizeOf();
         // for types of unknown size int
-        it = typeSizeLookup.insert(tn, size ? size : sizeof(int));
+        it = typeSizeLookup.insert(tn, size ? size_t(size) : sizeof(int));
     }
     return *it;
 }
